Enhancement to optionally post (tweet) media file from RSS feed on Twitter.
If the optional TweepyClient "post_include_media" is set to true, an attempt will be made to include a
media file contained in the source RSS feed.  This entails first downloading the media file (to an
optionally specified $MEDIA_DIR location, otherwise /tmp), and then including the resolved local path
to the downloaded file as part of the alternate Tweepy API for posting purposes.
Modifications have also been made to the Dockerfile to include and specify $MEDIA_DIR as a subdirectory
of $DATA_DIR.