Enhancement to optionally include RSS feed content withing Twitter postings (tweets).
When enabled by the optional TweepyClient post_include_content setting of true, the content (or
description, if content is empty) of a given RSS feed entry will be included as part of the tweet.
Any hashtags encountered at the end of this content/description will be removed for processing as
keywords.  Content pre-processing also includes removal of all HTML tags/constructs so that only
raw text remains.