Allow update/repost of already-published items
If an item's information includes the date upon which it was published, this information
is used as part of the "published" database entry, along with the relevant link.  This
allows an item with the same link to be published again later, as long as its published
date is different.  For those items containing no publication date, the former behavior
is unchanged.
Only the RSS feed implementation for this feature is in place at this point, but it would
be trivial to implement this functionality for Atom feeds as well.