Modifications to improve testability
Added optional command-line argument -t/--testing to indicate a test-only run, which
generates client-specific text output instead of posting to the respective client.
A value can also be provided to this argument, which will influence the default
configuration file name (feedspora.yml) and published DB name (feedspora.db).
If running in this test mode, no connections to the clients are made, setting the
stage for automated testing of all clients without the security risk associated with
client API keys or other such information to be present.